[ADD] return

+++ database/migrations/2024_10_05_075330_create_car_returns_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId("rental_id")->references('id')->on('rentals')->onUpdate('cascade')->onDelete('cascade');
             $table->date('return_date');
             $table->integer('total_days');
-            $table->decimal('total_cost', 10, 2);
+            $table->double('total_cost');
             $table->timestamps();
         });
     }
